Port changes from basho/riak into node_package
Many bug fixes and useability changes have happened in Riak
that needed to be copied over to node_package. The major change
was to specify `/bin/sh` as the shell to enable node_package
use on systems which don't ship with `bash` namely Solaris.

In addition, node_package used LOGNAME in its scripts which
was found to be unreliable across platforms in Riak. Changes
were made to use whois instead to fix #10.

Fixes: #10 #12