Add support for SmartOS 13.1
This is a major cleanup of the Makefile to determine what
dependencies are needed as well as determining the
SMF manifest file to use.

In addition, directories that are found in the rel/<appname>/data
directory are now automatically installed on SmartOS's
platform_data_dir.  This fixes previous bugs where Riak's
SNMP would crash when it couldn't access its database
directory.