remove contract specification from SMF manifests (solaris)
SMF uses the property, that this commit removes to determine how
the service should be affected under failure conditions.
This commit makes it so SMF reverts to its default behavior
to consider a service as "failed" if *any* of the processes
that make up the service dump core, or if *any* of the processes
receive a fatal signal originating outside of the service (ie.
if a processes raises its own signal or sends a signal to another
process within the same service that doesn't cause a
core dump, SMF ignores it)

It is in my experience running Riak in production at Voxer, if
a core dump is generated, the service will think riak is alive and
healthy, even though it is completely unresponsive.

This change makes it so any fatal signal to any of the processes
controlled by SMF for riak will trigger SMF to restart the service,
as well as if any of the processes dump core