Fix issue with dpkg purges failing in the postrm script
The postrm causes a purge to fail to run to completion if either the
pkill command fails to kill any processes or the group deletion fails
because the group is not empty. These are acceptable conditions and
should not cause the purge to fail Change each respective command so
that failure does not cause the postrm script to return an error code
and erroneously cause failure of the purge.