Apply rpm version information to app_version filed in vars.config
This applies the _revision field in the RPM spec file to the
app_version template variable.  This fixes the issue where
RPMs would not show information for the `runner version` command
due to the env.sh not having the proper variables filled out.

The PKG_VERSION makefile variable is set to the _revision
variable which now gets added to the vars.config properly.

Fixes: basho/node_package#69