Conditionally include manpages in file check for RPMs
If manpages did not exist for any of the `package_commands` list
RPM would fail when adding the {_mandir}/man1/* directive to
the %files section of the spec file.  This fix dynamically adds
each manpage to an `additional_files_list` that gets read
in by the %files section later on.