Change rpm spec file to setup on the package_install_name
This is a conflict between how riak-cs-ee and riak_ee are structured.
riak-cs-ee is built from the riak-cs repo, with changes in the
rebar.config.  They share the same $(PKGID) `riak-cs` so the
resulting tarname is riak-cs-1.4.0.tar.gz and the unpacked directory
is also riak-cs-1.4.0.  riak-cs-ee shares these settings, so
if %prep looks for riak-cs-ee-1.4.0 it will fail.  So with this
commit riak-cs-ee will work, but riak_ee will not.  This needs
to be fixed in riak_ee to make for a possible solution between
both products.