Replace sudo in runner scripts with su for greater compatibility
Fixes: basho/node_package#125

This has been a much needed change for a long time.  Requiring
sudo made for unhappy admin's and difficulty supporting all
OS's equally.  This also removes the need to install sudo on
OS's that don't typically ship with it, FreeBSD, Solaris, etc.

In addition this fix makes the `tail` command used several times
so that it works on solaris without using xpg4's tail.