Change nodetool to permit zero parameter rpc calls
Change nodetool so that functions of zero arity may be called using
the supplied rpc facility. This is done by making a distinction
between the case when rpc parameters are present and when they are
not.

Previously calling zero arity functions has not been possible because
even submitting no rpc parameters is treated the same as submitting
the request with a single parameter of an empty list.  Nodetool
encapsulates any parameters as a list and the result of attempting to
call a function that takes no parameters was the rpc function calling
the supplied function with a single parameter of an empty list.

This change does not preclude calling functions with a single
parameter of an empty list. It just now makes that decision explicit
on the part of the caller.