Handle net_ticktime argument specially. Resolves RIAK-1281
`escript` does not process emulator/runtime arguments like `erl`, so you
can't blindly pass an application environment setting on the command
line. This meant that the `-kernel net_ticktime N` argument passed to
nodetool was not functioning as intended and being treated as regular
arguments, resulting in a crash.

The net_ticktime setting is also a sensitive setting and should likely
be set before the Erlang distribution is started. I can only assume that
the original author that added it had intended for it to be loaded by
the application_controller, as `erl` does.

Therefore, this change handles net_ticktime specially in
nodetool:process_args/3 and puts it as the first argument to the script
so that it is set first.