Set NODETOOL up as an alias and include ERL_FLAGS env vars
nodetool won't work with TLS distribution as the erlang vm args
don't get passed in to escript. This commit reads the TLS args from
the vm.args file if it exists and prefixes the call to escript with
them.
The only way to have this work properly is to alias the command rather
than store it as a variable

basho/riak#509