Ensure proper perms on all SMF files for SmartOS
I'm not entirely clear on why this is necessary, because it seems like
the @owner/@group directives that the Makefile generates should take
care of it for us. But this should hopefully fix the issue that Doug and
I saw where the SMF files were getting installed as the user that built
the package instead of root.