Fix use of 'id' on Solaris 10
Commit 8d556bf0860fc628d273822dcf5fb063b107b3a1 changed
how the current user is determined from using 'whoami' to
'id'. On Solaris10 only the version of 'id' in '/usr/xpg4/bin'
knows about the '-u' flag.

This fix simply changes the PATH variable to use
'/usr/xpg4/bin' instead of '/usr/ucb', which was previously
added to support 'whoami'.