Use POSIX compatible shell and tools on Solaris
The runner (a.k.a riak, riak-cs, etc) script sourced
env.sh early on, before pivoting to running in a POSIX
compliant shell. Pivoting first, and sourcing env.sh afterwards
allows us to write the remainder of the script and any sourced
scripts in a POSIX compliant way without breaking execution on
Solaris 10.