Allow check_pid_status to differentiate similar names (#216)
Without a trailing space in the grep command, two different programs
with the same prefix (e.g. riak and riak-cs) will be confused. In
this example, if riak and riak-cs are configured to run on the same
server the riak init script will incorrectly say that it is running
in the situation where riak is stopped and riak-cs is running.

Also, compare to the grep line in hardstop (which already has the
trailing space).